change f32 to f64 everywhere

+++ src/main.rs
@@ -6,7 +6,7 @@ use nalgebra::Vector3;
 use ray::Ray;
 mod ray;
 
-fn hit_sphere(center: Vector3<f32>, radius: f32, ray: &Ray) -> bool {
+fn hit_sphere(center: Vector3<f64>, radius: f64, ray: &Ray) -> bool {
     let oc = ray.origin - center;
     let a = ray.direction.dot(&ray.direction);
     let b = 2.0 * oc.dot(&ray.direction);
@@ -35,14 +35,14 @@ fn get_color(ray: Ray) -> Vector3<u8> {
     return float_color_to_byte_color(new_color)
 }
 
-fn float_color_to_byte_color(color: Vector3<f32>) -> Vector3<u8> {
+fn float_color_to_byte_color(color: Vector3<f64>) -> Vector3<u8> {
     let result = color
     .iter()
     .map(|x| float_pixel_to_byte(x));
     return Vector3::from_iterator(result);
 }
 
-fn float_pixel_to_byte(pixel: &f32) -> u8 {
+fn float_pixel_to_byte(pixel: &f64) -> u8 {
     let scaled = pixel * 255.0;
     if scaled > 255.0 {
         return 255 
@@ -53,20 +53,20 @@ fn float_pixel_to_byte(pixel: &f32) -> u8 {
 }
 
 fn main() {
-    let aspect_ratio: f32 = 16.0 / 9.0;
+    let aspect_ratio: f64 = 16.0 / 9.0;
     let image_width: usize = 400;
-    let image_height: usize = (image_width as f32 / aspect_ratio) as usize;
-    let view_height : f32 = 2.0;
-    let view_width: f32 = view_height * (image_width as f32 / image_height as f32);
+    let image_height: usize = (image_width as f64 / aspect_ratio) as usize;
+    let view_height : f64 = 2.0;
+    let view_width: f64 = view_height * (image_width as f64 / image_height as f64);
 
-    let focal_length: f32 = 1.0;
+    let focal_length: f64 = 1.0;
     let camera_point = Vector3::new(0.0,0.0,0.0);
 
     let viewport_u = Vector3::new(view_width,0.0,0.0);
     let viewport_v = Vector3::new(0.0,-view_height,0.0);
 
-    let pixel_delta_u = viewport_u / (image_width as f32);
-    let pixel_delta_v = viewport_v / (image_height as f32);
+    let pixel_delta_u = viewport_u / (image_width as f64);
+    let pixel_delta_v = viewport_v / (image_height as f64);
 
     let view_upper_left = camera_point - Vector3::new(0.0,0.0,focal_length) - viewport_u/2.0 - viewport_v/2.0;
 
@@ -79,7 +79,7 @@ fn main() {
     for j in 0..image_height {
         for i in 0..image_width {
             bar.inc(1);
-            let pixel_center = pixel00_loc + (i as f32 * pixel_delta_u) + (j as f32 * pixel_delta_v);
+            let pixel_center = pixel00_loc + (i as f64 * pixel_delta_u) + (j as f64 * pixel_delta_v);
             let ray_direction = pixel_center - camera_point;
             let r = Ray::new(camera_point,ray_direction);
             let color = get_color(r);
