+++ src/ray.rs
@@ -1,16 +1,16 @@
 use nalgebra::Vector3;
 
 pub struct Ray {
-    pub origin: Vector3<f32>,
-    pub direction: Vector3<f32>,
+    pub origin: Vector3<f64>,
+    pub direction: Vector3<f64>,
 }
 
 impl Ray {
-    pub fn new(origin: Vector3<f32>, direction: Vector3<f32>) -> Self {
+    pub fn new(origin: Vector3<f64>, direction: Vector3<f64>) -> Self {
         Self { origin, direction }
     }
 
-    pub fn at(&self, t: f32) -> Vector3<f32> {
+    pub fn at(&self, t: f64) -> Vector3<f64> {
         self.origin + t * self.direction
     }
 }
